requirements.txt
requests
beautifulsoup4
feedparser
openai
python-dotenv
lxml
pyyaml
